+++ app/src/main/java/com/patloew/oeffitracker/data/model/TicketStatistics.kt
@@ -6,14 +6,14 @@ import java.time.LocalDate
 data class TicketWithStatistics(
     val id: Long,
     val name: String,
-    val price: Int,
-    val deduction: Int?,
+    val price: Long,
+    val deduction: Long?,
     val startDate: LocalDate,
     val endDate: LocalDate,
     val createdTimestamp: Long,
-    val fareSum: Int,
-    val additionalCostsSum: Int?,
+    val fareSum: Long,
+    val additionalCostsSum: Long?,
     val durationSum: Duration?,
     val delaySum: Duration?,
-    val distanceSum: Float?
+    val distanceSum: Double?
 )
